Job Cloud/cloud_run_b5/requirements.txt
Flask==3.0.3
google-cloud-bigquery==3.25.0
